Move service to resolvable location
